+++ prettier.config.mjs
@@ -1,7 +1,0 @@
-export default {
-  arrowParens: "avoid",
-  endOfLine: "lf",
-  semi: true,
-  singleQuote: true,
-  trailingComma: "all",
-};